Users can now post their highscores. Highscore limit per game mode is currently 5 for testing purposes.

+++ backend/migrations/20230314134152-create-highscores-table.js
@@ -9,7 +9,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       user_id: {
-        type: Sequelize.SMALLINT
+        type: Sequelize.INTEGER
       },
       user_name: {
         allowNull: false,
